refactor: update and organize types
