+++ src/services/code-index/interfaces/embedder.ts
@@ -1,14 +1,21 @@
-import { EmbeddingResponse } from "./types"
-
 /**
- * Interface for embedding providers
+ * Interface for code index embedders.
+ * This interface is implemented by both OpenAI and Ollama embedders.
  */
 export interface IEmbedder {
 	/**
-	 * Creates embeddings for the given texts
-	 * @param texts Array of text strings to embed
-	 * @param model Optional model identifier
-	 * @returns Promise resolving to embedding response
+	 * Creates embeddings for the given texts.
+	 * @param texts Array of text strings to create embeddings for
+	 * @param model Optional model ID to use for embeddings
+	 * @returns Promise resolving to an EmbeddingResponse
 	 */
 	createEmbeddings(texts: string[], model?: string): Promise<EmbeddingResponse>
+}
+
+export interface EmbeddingResponse {
+	embeddings: number[][]
+	usage?: {
+		promptTokens: number
+		totalTokens: number
+	}
 }
