+++ src/services/code-index/interfaces/index.ts
@@ -1,4 +1,3 @@
-export * from "./types"
 export * from "./embedder"
 export * from "./vector-store"
 export * from "./file-processor"
